Separate compilers and intermediary language ASTs
