+++ intlc.cabal
@@ -45,9 +45,11 @@ library
   exposed-modules:
     Intlc.Compiler
     Intlc.Backend.Common
-    Intlc.Backend.JavaScript
-    Intlc.Backend.ICU
-    Intlc.Backend.TypeScript
+    Intlc.Backend.JavaScript.Language
+    Intlc.Backend.JavaScript.Compiler
+    Intlc.Backend.ICU.Compiler
+    Intlc.Backend.TypeScript.Language
+    Intlc.Backend.TypeScript.Compiler
     Intlc.Core
     Intlc.ICU
     Intlc.Parser
